test(learning): split qe-reasoning-bank.test.ts into three forks (issue #448 step 2)
The original 1019-line file ran 58 tests in a single vitest fork, each
spinning up a full QEReasoningBank (transformer embeddings + HNSW + WASM
coherence). Cumulative per-test memory growth blew past the heap ceiling
and the publish workflow timed out on multiple v3.9.x releases.

This commit extracts the file into three forks so the heap pressure is
distributed across three sequential processes (vitest config:
pool=forks, maxForks=1, fileParallelism=false):

  - qe-reasoning-bank-utilities.test.ts (NEW, 29 tests, no bank)
    Pure-function tests for qe-patterns + qe-guidance. Module-graph weight
    is negligible — runs in ~10ms with no transformer load. Safe to
    include in test:unit:fast.

  - qe-reasoning-bank-feedback-loop.test.ts (NEW, 9 tests, full bank)
    recordOutcome / qe_pattern_usage feedback-loop tests with vi.doMock
    churn over '../../../src/kernel/unified-memory.js'. Carries the
    heaviest mocking surface, so isolated to its own fork.

  - qe-reasoning-bank.test.ts (SLIMMED, 20 tests, full bank)
    Init / Pattern Storage / Task Routing / Learning Outcomes /
    Guidance Generation / Statistics / Embedding. Header comment
    updated to reference the two companion files.

Both heavy files (-feedback-loop and the slim core) remain excluded from
test:unit:fast for now. Re-enabling them is acceptance #2 of the issue
and requires step 3 of the proposed solution (lazy / disposable hooks on
real-embeddings.ts and coherence-service.ts) — local repro in a 16GB
Codespace shows even the slim 9-test fork still OOMs at 6GB heap,
indicating per-test cumulative growth that file-splitting alone doesn't
cure. The split is preparatory and reduces the leak slope from
58×Δ to max(20,9,0)×Δ per fork, which should give CI headroom and
clear the way for the dispose-hooks fix in step 3.

Issue: #448
