+++ tests/unit/learning/qe-reasoning-bank.test.ts
@@ -1,20 +1,20 @@
 /**
- * Unit Tests - QE ReasoningBank
+ * Unit Tests - QE ReasoningBank (core suite)
  * ADR-021: QE ReasoningBank for Pattern Learning
  *
- * MEMORY NOTE: This file is intentionally excluded from `npm run test:unit:fast`
- * (see package.json). It spins up a full reasoning-bank with real
+ * MEMORY NOTE: This file spins up a full reasoning-bank with real
  * transformer embeddings (`@xenova/transformers`, all-MiniLM-L6-v2) and
- * HNSW indexing on every test. 58 tests × full bank init = cumulative
- * memory growth that exceeds an 8GB codespace's default heap.
+ * HNSW indexing on every test, so it's run in its own vitest fork
+ * (pool=forks, maxForks=1, fileParallelism=false). Two companion files
+ * carry the heaviest sub-suites out of this fork (issue #448, step 2):
  *
- * This is NOT a flaky test in the "broken" sense — it's resource-heavy.
- * Run via:
- *   - `npm run test:ci` (full suite — CI has the headroom)
- *   - `NODE_OPTIONS='--max-old-space-size=4096' npx vitest run tests/unit/learning/qe-reasoning-bank.test.ts` (locally, if needed)
+ *   - qe-reasoning-bank-feedback-loop.test.ts — recordOutcome /
+ *     qe_pattern_usage feedback-loop tests with vi.doMock churn.
+ *   - qe-reasoning-bank-utilities.test.ts — pure-function tests for
+ *     qe-patterns + qe-guidance that don't need a bank at all.
  *
- * If a true memory leak ever surfaces here (single test OOMs in CI), the
- * suspects are: the transformer pipeline module-singleton in
+ * If a true memory leak ever surfaces here (single test OOMs in CI),
+ * the suspects are: the transformer pipeline module-singleton in
  * real-embeddings.ts, accumulated HNSW vectors that aren't released by
  * patternStore.dispose(), or pretrained pattern bundles retained beyond
  * the test boundary. Audit those before declaring this file flaky.
@@ -31,24 +31,7 @@ import {
 // Ensure these tests exercise the in-memory PatternStore, not the RVF variant
 beforeEach(() => { setRuVectorFeatureFlags({ useRVFPatternStore: false }); });
 afterEach(() => { resetRuVectorFeatureFlags(); });
-import {
-  detectQEDomain,
-  detectQEDomains,
-  mapQEDomainToAQE,
-  calculateQualityScore,
-  shouldPromotePattern,
-  validateQEPattern,
-  applyPatternTemplate,
-  QEPattern,
-  QEPatternType,
-  QEDomain,
-} from '../../../src/learning/qe-patterns';
-import {
-  getGuidance,
-  getCombinedGuidance,
-  generateGuidanceContext,
-  checkAntiPatterns,
-} from '../../../src/learning/qe-guidance';
+
 import { createMockMemory } from '../../mocks';
 import type { MemoryBackend } from '../../../src/kernel/interfaces';
 import { checkRuvectorPackagesAvailable } from '../../../src/integrations/ruvector/wrappers';
@@ -309,285 +292,6 @@ describe.runIf(canTest.gnn)('QE ReasoningBank', () => {
     });
   });
 
-  describe('recordOutcome - qe_pattern_usage feedback loop', () => {
-    // Helper: store a pattern and return its ID
-    async function storeTestPattern(): Promise<string | null> {
-      const storeResult = await reasoningBank.storePattern({
-        patternType: 'test-template',
-        name: 'Analytics Feedback Pattern',
-        description: 'Pattern to test qe_pattern_usage INSERT',
-        template: { type: 'code', content: 'test content', variables: [] },
-      });
-      if (!storeResult.success) return null;
-      return storeResult.value.id;
-    }
-
-    it('should write to qe_pattern_usage when unified memory is available', async () => {
-      const patternId = await storeTestPattern();
-      expect(patternId).not.toBeNull();
-      if (!patternId) return;
-
-      // Create a mock database with a prepare().run() chain
-      const mockRun = vi.fn();
-      const mockPrepare = vi.fn().mockReturnValue({ run: mockRun });
-      const mockDb = { prepare: mockPrepare };
-      const mockGetUnifiedMemory = vi.fn().mockReturnValue({
-        getDatabase: () => mockDb,
-      });
-
-      // Mock the dynamic import to return our mock
-      const originalImport = await import('../../../src/kernel/unified-memory.js').catch(() => null);
-      vi.doMock('../../../src/kernel/unified-memory.js', () => ({
-        ...originalImport,
-        getUnifiedMemory: mockGetUnifiedMemory,
-      }));
-
-      const outcomeResult = await reasoningBank.recordOutcome({
-        patternId,
-        success: true,
-        metrics: { testsPassed: 10, testsFailed: 0 },
-        feedback: 'Excellent coverage',
-      });
-
-      expect(outcomeResult.success).toBe(true);
-
-      // The patternStore.recordUsage() should still work
-      const pattern = await reasoningBank.getPattern(patternId);
-      expect(pattern?.usageCount).toBe(1);
-      expect(pattern?.successfulUses).toBe(1);
-
-      // Restore mock so other tests are unaffected
-      vi.doUnmock('../../../src/kernel/unified-memory.js');
-    });
-
-    it('should not fail when unified memory is NOT available', async () => {
-      const patternId = await storeTestPattern();
-      expect(patternId).not.toBeNull();
-      if (!patternId) return;
-
-      // Mock the dynamic import to throw (simulating unified memory not initialized)
-      vi.doMock('../../../src/kernel/unified-memory.js', () => {
-        throw new Error('Unified memory not initialized');
-      });
-
-      // recordOutcome should succeed despite the analytics INSERT failing
-      const outcomeResult = await reasoningBank.recordOutcome({
-        patternId,
-        success: true,
-        metrics: { testsPassed: 5, testsFailed: 0 },
-      });
-
-      expect(outcomeResult.success).toBe(true);
-
-      // The patternStore.recordUsage() should still have worked
-      const pattern = await reasoningBank.getPattern(patternId);
-      expect(pattern?.usageCount).toBe(1);
-      expect(pattern?.successfulUses).toBe(1);
-
-      vi.doUnmock('../../../src/kernel/unified-memory.js');
-    });
-
-    it('should not fail when getDatabase() throws', async () => {
-      const patternId = await storeTestPattern();
-      expect(patternId).not.toBeNull();
-      if (!patternId) return;
-
-      // Mock where getUnifiedMemory succeeds but getDatabase throws
-      vi.doMock('../../../src/kernel/unified-memory.js', () => ({
-        getUnifiedMemory: () => ({
-          getDatabase: () => { throw new Error('Database not ready'); },
-        }),
-      }));
-
-      const outcomeResult = await reasoningBank.recordOutcome({
-        patternId,
-        success: false,
-        metrics: { testsPassed: 1, testsFailed: 9 },
-        feedback: 'Pattern produced poor results',
-      });
-
-      expect(outcomeResult.success).toBe(true);
-
-      // patternStore should still have recorded the failure
-      const pattern = await reasoningBank.getPattern(patternId);
-      expect(pattern?.usageCount).toBe(1);
-      expect(pattern?.successfulUses).toBe(0);
-
-      vi.doUnmock('../../../src/kernel/unified-memory.js');
-    });
-
-    it('should not fail when db.prepare().run() throws', async () => {
-      const patternId = await storeTestPattern();
-      expect(patternId).not.toBeNull();
-      if (!patternId) return;
-
-      // Mock where prepare().run() throws (e.g., table does not exist)
-      vi.doMock('../../../src/kernel/unified-memory.js', () => ({
-        getUnifiedMemory: () => ({
-          getDatabase: () => ({
-            prepare: () => ({
-              run: () => { throw new Error('no such table: qe_pattern_usage'); },
-            }),
-          }),
-        }),
-      }));
-
-      const outcomeResult = await reasoningBank.recordOutcome({
-        patternId,
-        success: true,
-        metrics: { testsPassed: 8, testsFailed: 2 },
-      });
-
-      expect(outcomeResult.success).toBe(true);
-
-      const pattern = await reasoningBank.getPattern(patternId);
-      expect(pattern?.usageCount).toBe(1);
-      expect(pattern?.successfulUses).toBe(1);
-
-      vi.doUnmock('../../../src/kernel/unified-memory.js');
-    });
-
-    it('should record success=true outcome and increment stats correctly', async () => {
-      const patternId = await storeTestPattern();
-      expect(patternId).not.toBeNull();
-      if (!patternId) return;
-
-      const statsBefore = await reasoningBank.getStats();
-      const outcomesBefore = statsBefore.learningOutcomes;
-
-      const outcomeResult = await reasoningBank.recordOutcome({
-        patternId,
-        success: true,
-        metrics: { testsPassed: 15, testsFailed: 0, coverageImprovement: 12.5 },
-        feedback: 'All tests green',
-      });
-
-      expect(outcomeResult.success).toBe(true);
-
-      const statsAfter = await reasoningBank.getStats();
-      expect(statsAfter.learningOutcomes).toBe(outcomesBefore + 1);
-
-      const pattern = await reasoningBank.getPattern(patternId);
-      expect(pattern?.usageCount).toBe(1);
-      expect(pattern?.successfulUses).toBe(1);
-    });
-
-    it('should record success=false outcome and not increment successfulOutcomes', async () => {
-      const patternId = await storeTestPattern();
-      expect(patternId).not.toBeNull();
-      if (!patternId) return;
-
-      const statsBefore = await reasoningBank.getStats();
-      const outcomesBefore = statsBefore.learningOutcomes;
-
-      const outcomeResult = await reasoningBank.recordOutcome({
-        patternId,
-        success: false,
-        metrics: { testsPassed: 0, testsFailed: 10, executionTimeMs: 5000 },
-        feedback: 'All tests failed due to stale mocks',
-      });
-
-      expect(outcomeResult.success).toBe(true);
-
-      const statsAfter = await reasoningBank.getStats();
-      expect(statsAfter.learningOutcomes).toBe(outcomesBefore + 1);
-
-      const pattern = await reasoningBank.getPattern(patternId);
-      expect(pattern?.usageCount).toBe(1);
-      expect(pattern?.successfulUses).toBe(0);
-    });
-
-    it('should pass null metrics_json when outcome has no metrics', async () => {
-      const patternId = await storeTestPattern();
-      expect(patternId).not.toBeNull();
-      if (!patternId) return;
-
-      const mockRun = vi.fn();
-      const mockPrepare = vi.fn().mockReturnValue({ run: mockRun });
-      const mockDb = { prepare: mockPrepare };
-      vi.doMock('../../../src/kernel/unified-memory.js', () => ({
-        getUnifiedMemory: () => ({
-          getDatabase: () => mockDb,
-        }),
-      }));
-
-      await reasoningBank.recordOutcome({
-        patternId,
-        success: true,
-      });
-
-      // patternStore should still track usage
-      const pattern = await reasoningBank.getPattern(patternId);
-      expect(pattern?.usageCount).toBe(1);
-      expect(pattern?.successfulUses).toBe(1);
-
-      vi.doUnmock('../../../src/kernel/unified-memory.js');
-    });
-
-    it('should pass null feedback when outcome has no feedback', async () => {
-      const patternId = await storeTestPattern();
-      expect(patternId).not.toBeNull();
-      if (!patternId) return;
-
-      const mockRun = vi.fn();
-      const mockPrepare = vi.fn().mockReturnValue({ run: mockRun });
-      const mockDb = { prepare: mockPrepare };
-      vi.doMock('../../../src/kernel/unified-memory.js', () => ({
-        getUnifiedMemory: () => ({
-          getDatabase: () => mockDb,
-        }),
-      }));
-
-      await reasoningBank.recordOutcome({
-        patternId,
-        success: false,
-        metrics: { testsPassed: 3, testsFailed: 7 },
-      });
-
-      const pattern = await reasoningBank.getPattern(patternId);
-      expect(pattern?.usageCount).toBe(1);
-      expect(pattern?.successfulUses).toBe(0);
-
-      vi.doUnmock('../../../src/kernel/unified-memory.js');
-    });
-
-    it('should early-return ok when enableLearning is false', async () => {
-      const noLearnMemory = createMockMemory();
-      const noLearnBank = createQEReasoningBank(noLearnMemory, undefined, {
-        enableLearning: false,
-      });
-      await noLearnBank.initialize();
-
-      // Store a pattern first
-      const storeResult = await noLearnBank.storePattern({
-        patternType: 'test-template',
-        name: 'No-Learn Pattern',
-        description: 'Should not record outcome',
-        template: { type: 'code', content: 'test', variables: [] },
-      });
-
-      if (!storeResult.success) {
-        await noLearnBank.dispose();
-        return;
-      }
-
-      const outcomeResult = await noLearnBank.recordOutcome({
-        patternId: storeResult.value.id,
-        success: true,
-        metrics: { testsPassed: 5, testsFailed: 0 },
-      });
-
-      // Should return ok(undefined) without calling recordUsage or the INSERT
-      expect(outcomeResult.success).toBe(true);
-
-      // Pattern should NOT have updated usage stats since learning is disabled
-      const pattern = await noLearnBank.getPattern(storeResult.value.id);
-      expect(pattern?.usageCount).toBe(0);
-
-      await noLearnBank.dispose();
-    });
-  });
-
   describe('Guidance Generation', () => {
     it('should get guidance for test-generation domain', () => {
       const guidance = reasoningBank.getGuidance('test-generation');
@@ -677,343 +381,6 @@ describe.runIf(canTest.gnn)('QE ReasoningBank', () => {
 
       // Different text should produce different embeddings
       expect(embedding1).not.toEqual(embedding2);
-    });
-  });
-});
-
-describe('QE Pattern Utilities', () => {
-  describe('detectQEDomain', () => {
-    it('should detect test-generation domain', () => {
-      expect(detectQEDomain('Write tests for the UserService')).toBe('test-generation');
-      expect(detectQEDomain('describe("MyComponent")')).toBe('test-generation');
-      expect(detectQEDomain('it("should work")')).toBe('test-generation');
-    });
-
-    it('should detect coverage-analysis domain', () => {
-      expect(detectQEDomain('Analyze code coverage')).toBe('coverage-analysis');
-      expect(detectQEDomain('Find uncovered branches')).toBe('coverage-analysis');
-    });
-
-    it('should detect security-compliance domain', () => {
-      expect(detectQEDomain('Check for XSS vulnerabilities')).toBe('security-compliance');
-      expect(detectQEDomain('OWASP security scan')).toBe('security-compliance');
-    });
-
-    it('should detect visual-accessibility domain', () => {
-      // Use unique keywords that only match visual-accessibility
-      // percy and a11y are unique to this domain, avoid "regression" (defect-intelligence)
-      expect(detectQEDomain('percy snapshot')).toBe('visual-accessibility');
-      expect(detectQEDomain('a11y audit')).toBe('visual-accessibility');
-    });
-
-    it('should return null for unmatched text', () => {
-      expect(detectQEDomain('Hello world')).toBeNull();
-    });
-  });
-
-  describe('detectQEDomains', () => {
-    it('should detect multiple domains', () => {
-      const domains = detectQEDomains(
-        'Generate tests with coverage analysis for security-critical code'
-      );
-      expect(domains).toContain('test-generation');
-      expect(domains).toContain('coverage-analysis');
-    });
-  });
-
-  describe('mapQEDomainToAQE', () => {
-    it('should map QE domains to AQE domains (identity mapping since aligned)', () => {
-      // QEDomain and DomainName are now aligned - this is an identity mapping
-      expect(mapQEDomainToAQE('test-generation')).toBe('test-generation');
-      expect(mapQEDomainToAQE('coverage-analysis')).toBe('coverage-analysis');
-      expect(mapQEDomainToAQE('security-compliance')).toBe('security-compliance');
-      expect(mapQEDomainToAQE('visual-accessibility')).toBe('visual-accessibility');
-    });
-  });
-
-  describe('calculateQualityScore', () => {
-    it('should calculate quality score', () => {
-      const score = calculateQualityScore({
-        confidence: 0.8,
-        usageCount: 50,
-        successRate: 0.9,
-      });
-
-      expect(score).toBeGreaterThan(0);
-      expect(score).toBeLessThanOrEqual(1);
-    });
-
-    it('should weight success rate highest', () => {
-      const highSuccess = calculateQualityScore({
-        confidence: 0.5,
-        usageCount: 10,
-        successRate: 1.0,
-      });
-
-      const lowSuccess = calculateQualityScore({
-        confidence: 0.5,
-        usageCount: 10,
-        successRate: 0.2,
-      });
-
-      expect(highSuccess).toBeGreaterThan(lowSuccess);
-    });
-  });
-
-  describe('shouldPromotePattern', () => {
-    const basePattern: QEPattern = {
-      id: '1',
-      patternType: 'test-template',
-      qeDomain: 'test-generation',
-      domain: 'test-generation',
-      name: 'Test',
-      description: 'Test',
-      confidence: 0.7,
-      usageCount: 10,
-      successRate: 0.8,
-      qualityScore: 0.7,
-      context: { tags: [] },
-      template: { type: 'code', content: '', variables: [] },
-      tier: 'short-term',
-      createdAt: new Date(),
-      lastUsedAt: new Date(),
-      successfulUses: 5,
-    };
-
-    it('should promote pattern with 3+ successful uses', () => {
-      const result = shouldPromotePattern(basePattern);
-      expect(result.meetsUsageCriteria).toBe(true);
-      expect(result.meetsQualityCriteria).toBe(true);
-      expect(result.meetsCoherenceCriteria).toBe(true);
-      expect(result.blockReason).toBeUndefined();
-    });
-
-    it('should not promote already long-term patterns', () => {
-      const result = shouldPromotePattern({ ...basePattern, tier: 'long-term' });
-      expect(result.meetsUsageCriteria).toBe(false);
-      expect(result.blockReason).toBe('insufficient_usage');
-    });
-
-    it('should not promote patterns with low confidence', () => {
-      const result = shouldPromotePattern({ ...basePattern, confidence: 0.4 });
-      expect(result.meetsQualityCriteria).toBe(false);
-      expect(result.blockReason).toBe('low_quality');
-    });
-
-    it('should not promote patterns with low success rate', () => {
-      const result = shouldPromotePattern({ ...basePattern, successRate: 0.5 });
-      expect(result.meetsQualityCriteria).toBe(false);
-      expect(result.blockReason).toBe('low_quality');
-    });
-
-    it('should not promote patterns with few successful uses', () => {
-      const result = shouldPromotePattern({ ...basePattern, successfulUses: 2 });
-      expect(result.meetsUsageCriteria).toBe(false);
-      expect(result.blockReason).toBe('insufficient_usage');
-    });
-
-    it('should block promotion when coherence energy exceeds threshold', () => {
-      const result = shouldPromotePattern(basePattern, 0.5, 0.4);
-      expect(result.meetsUsageCriteria).toBe(true);
-      expect(result.meetsQualityCriteria).toBe(true);
-      expect(result.meetsCoherenceCriteria).toBe(false);
-      expect(result.blockReason).toBe('coherence_violation');
-    });
-
-    it('should allow promotion when coherence energy is below threshold', () => {
-      const result = shouldPromotePattern(basePattern, 0.3, 0.4);
-      expect(result.meetsUsageCriteria).toBe(true);
-      expect(result.meetsQualityCriteria).toBe(true);
-      expect(result.meetsCoherenceCriteria).toBe(true);
-      expect(result.blockReason).toBeUndefined();
-    });
-
-    it('should allow promotion when coherence energy is not provided', () => {
-      const result = shouldPromotePattern(basePattern);
-      expect(result.meetsUsageCriteria).toBe(true);
-      expect(result.meetsQualityCriteria).toBe(true);
-      expect(result.meetsCoherenceCriteria).toBe(true);
-      expect(result.blockReason).toBeUndefined();
-    });
-  });
-
-  describe('validateQEPattern', () => {
-    it('should validate valid pattern', () => {
-      const pattern: Partial<QEPattern> = {
-        id: '1',
-        patternType: 'test-template',
-        qeDomain: 'test-generation',
-        name: 'Test Pattern',
-        template: {
-          type: 'code',
-          content: 'describe("{{name}}")',
-          variables: [{ name: 'name', type: 'string', required: true }],
-        },
-        confidence: 0.8,
-        successRate: 0.9,
-      };
-
-      const result = validateQEPattern(pattern);
-      expect(result.valid).toBe(true);
-      expect(result.errors).toHaveLength(0);
-    });
-
-    it('should detect missing required fields', () => {
-      const result = validateQEPattern({});
-      expect(result.valid).toBe(false);
-      expect(result.errors.length).toBeGreaterThan(0);
-    });
-
-    it('should detect invalid confidence range', () => {
-      const result = validateQEPattern({
-        id: '1',
-        patternType: 'test-template',
-        qeDomain: 'test-generation',
-        name: 'Test',
-        template: { type: 'code', content: 'test', variables: [] },
-        confidence: 1.5, // Invalid
-      });
-
-      expect(result.valid).toBe(false);
-      expect(result.errors).toContain('Confidence must be between 0 and 1');
-    });
-  });
-
-  describe('applyPatternTemplate', () => {
-    it('should apply template variables', () => {
-      const template = {
-        type: 'code' as const,
-        content: 'describe("{{className}}", () => { {{testBody}} });',
-        variables: [
-          { name: 'className', type: 'string' as const, required: true },
-          { name: 'testBody', type: 'code' as const, required: true },
-        ],
-      };
-
-      const result = applyPatternTemplate(template, {
-        className: 'UserService',
-        testBody: 'it("should work", () => {});',
-      });
-
-      expect(result).toContain('UserService');
-      expect(result).toContain('should work');
-    });
-
-    it('should use default values', () => {
-      const template = {
-        type: 'code' as const,
-        content: 'describe("{{name}}", () => { {{async}} });',
-        variables: [
-          { name: 'name', type: 'string' as const, required: true },
-          { name: 'async', type: 'string' as const, required: false, defaultValue: '// tests' },
-        ],
-      };
-
-      const result = applyPatternTemplate(template, { name: 'Test' });
-      expect(result).toContain('// tests');
-    });
-
-    it('should throw for missing required variables', () => {
-      const template = {
-        type: 'code' as const,
-        content: '{{required}}',
-        variables: [{ name: 'required', type: 'string' as const, required: true }],
-      };
-
-      expect(() => applyPatternTemplate(template, {})).toThrow(
-        'Required variable required not provided'
-      );
-    });
-  });
-});
-
-describe('QE Guidance', () => {
-  describe('getGuidance', () => {
-    it('should return guidance for all domains', () => {
-      // Use the 12 DDD bounded context domains
-      const domains: QEDomain[] = [
-        'test-generation',
-        'test-execution',
-        'coverage-analysis',
-        'quality-assessment',
-        'defect-intelligence',
-        'requirements-validation',
-        'code-intelligence',
-        'security-compliance',
-        'contract-testing',
-        'visual-accessibility',
-        'chaos-resilience',
-        'learning-optimization',
-      ];
-
-      for (const domain of domains) {
-        const guidance = getGuidance(domain);
-        expect(guidance).toBeDefined();
-        // Note: guidance.domain may differ from input domain as registry reuses guidance objects
-        // (e.g., test-execution uses test-generation guidance)
-        expect(guidance.domain).toBeDefined();
-        expect(guidance.bestPractices.length).toBeGreaterThan(0);
-      }
-    });
-  });
-
-  describe('getCombinedGuidance', () => {
-    it('should combine domain, framework, and language guidance', () => {
-      const guidance = getCombinedGuidance('test-generation', {
-        framework: 'vitest',
-        language: 'typescript',
-      });
-
-      expect(guidance.length).toBeGreaterThan(0);
-      expect(guidance.some((g) => g.includes('[vitest]'))).toBe(true);
-      expect(guidance.some((g) => g.includes('[typescript]'))).toBe(true);
-    });
-
-    it('should include anti-patterns when requested', () => {
-      const guidance = getCombinedGuidance('test-generation', {
-        includeAntiPatterns: true,
-      });
-
-      expect(guidance.some((g) => g.includes('[AVOID]'))).toBe(true);
-    });
-  });
-
-  describe('generateGuidanceContext', () => {
-    it('should generate markdown context', () => {
-      const context = generateGuidanceContext('test-generation', {
-        framework: 'vitest',
-        language: 'typescript',
-      });
-
-      expect(context).toContain('## QE Guidance');
-      expect(context).toContain('Best Practices');
-      expect(context).toContain('Anti-Patterns');
-    });
-  });
-
-  describe('checkAntiPatterns', () => {
-    it('should detect flaky assertion patterns', () => {
-      const code = `
-        it('should be fast', async () => {
-          await setTimeout(() => {}, 100);
-          expect(Date.now()).toBeLessThan(Date.now() + 1000);
-        });
-      `;
-
-      const antiPatterns = checkAntiPatterns('test-generation', code);
-      expect(antiPatterns.some((ap) => ap.name === 'Flaky Assertion')).toBe(true);
-    });
-
-    it('should return empty for clean code', () => {
-      const code = `
-        it('should add numbers', () => {
-          expect(1 + 1).toBe(2);
-        });
-      `;
-
-      const antiPatterns = checkAntiPatterns('test-generation', code);
-      // May or may not detect patterns depending on implementation
-      expect(Array.isArray(antiPatterns)).toBe(true);
     });
   });
 });
